Rename electrodes to channels
This helps to clarify the relationship between channels and
electrodes.  Each relay switch (i.e. channel) can control one or
more electrodes (i.e. bussing).  It is possible to have multiple
channels hooked up to a single (or group of) electrode(s), which
could lead to a short circuit condition if the channels are in
different states, but it may be useful for resistive heating.