Fix autoconnection of DmfControlBoard and function generator
The function generator class was opening and blocking the serial
port connected to the DmfControlBoard, with the result that the
DmfControlBoard could not connect.  This change properly closes
the COM port if it is not a proper GPIB instrument.

When connecting to the DmfControlBoard, it also checks that the
name and hardware version match known values.