Expose Arduino constants and set default pin mode and state
Expose Arduino constants INPUT, OUTPUT, HIGH and LOW in the
dmf_control_board module namespace. Also, add functions for
checking and setting the default pin mode and pin state
of the digital pins on the Arduino. These settings are stored in
the EEPROM and restored as soon as the Arduino boots up.