Initial implementation of a plugin framework
Refs Ticket #4

This commit indroduces basic foundations for a plugin component
architecture based on the PyUtilib module.  Currently, any class
defined in a *.py file within the "plugins" directory and
implementing the IPlugin interface is automatically loaded and
enabled.

The biggest remaining limitation is that the state of plugins is
not yet stored when a protocol is saved.