Support saving state for plugins
Add 2 new functions to the IPlugin interface: on_protocol_load
and on_protocol_save. When on_protocol_save() is called, plugins
can pickle any state information to a string and add it to the
plugin_data dictionary stored by the Protocol object. When a
protocol is loaded, the on_protocol_load() function is called for
each plugin with the version number and pickled state as
parameters.

Closes #4.