Add Experiment Log viewer and convert controllers to plugins
The new Experiment Log Viewer is accessible from the "View" menu and it gives
details on previous experiments.

This commit also converts most of the controller classes into plugins. A new
function, emit_signals, has also been added to the plugin_manager class
which makes it easier for objects to send signals to other plugins.