Add version checking to plugin install, see #28
Read yaml encoded dictionary of the form:
    {'name': <plugin module name>, 'version': <string Version>}
from file named "properties.yml" in the plugin directory to obtain version of
plugin to be installed and from the currently installed version (if present).

The following is an example plugin directory layout an plugin archive:
    video_recorder.tgz
        \->video_recorder
            \->__init__.py
            \->microdrop
                \->__init__.py
            \->properties.yml

Example contents of the properties.yml file are shown below:
$ cat plugins/video_recorder/properties.yml
{name: video_recorder, version: '0.1'}

If a plugin is installed/upgraded inform the user that the application needs to
be restarted, then quit the application.

These changes address proposed change (2)-(3) from ticket #28.