Modify processing of config file
The config file is now processed (and the app options set) by the
on_app_init handler of the config_controller class. If other plugins
want to be able to access their app options from their on_app_init
handler, they should request that the config_controller is scheduled
before them.

This commit also makes the registration matrix an app option so
that it is persistent when the progam is restarted.

Fixes #76.