[SUBMODULE] Add support for mappers in protocol...
[SUBMODULE] Add support for pygtkhelpers PropertyMapper lists as a Fields
item attribute.  Using this, it is possible to change the state of a cell in
the protocol grid view based on program state.  For example, as of this commit,
many of the feedback fields from the DMF control board plugin are disabled when
feedback is disabled for the corresponding step.
Fixes #69.