Update GUI re: #80, #81
Fixes #80 by emphasizing that the Voltage, Frequency, Duration parameters relate
to the current step. Rename "duration" to "minimum duration" (#81) but to close #81,
we should also rename the duration member of DmfControlBoardOptions so that it
is also reflected in the protocol grid, or better yet, pull the duration
parameter out of the dmf_control_board plugin since it is actually more general.