When upgrading a plugin, install new version to...
When upgrading a plugin, install new version to new path.  Use the original
plugin path and increment a count to append onto the end until a unique
directory name is found.  This allows for the new version of a plugin to be
installed before the existing plugin is deleted.  This behaviour is required
since the deletion of the existing plugin must be delayed until the program is
restarted (due to file locking).

Fixes #67 (confirmed on Ubuntu, need to test in Windows)