[SUBMODULE] Decouple dmf_control_board plugin
Fixes #89.
Convert DmfControlBoardPlugin to a Plugin.
Eliminate app.control_board and any references to it.
Skip updates that involve a device or experiment log if the corresponding
app attribute is not ready.

Note that the experiment log plugin and protocol controller refer to the
dmf_control_board plugin.  However, they should now run fine without the
dmf_control_board installed.