Modify strategy for setting default dialog parent
Instead of using a YesNo callable class with a default parent setting, add a
Defaults class instance named "DEFAULTS" to utility.gui.  The default parent to
use for dialogs can be set by assigning a parent widget to
DEFAULTS.parent_widget.  The benefit of this approach is that all utility.gui
code can make use this default parent value.  As part of this commit, all
dialogs in utility.gui are set to use DEFAULTS.parent_widget if no parent was
explicitly specified.

Fixes #98.