Fix #86, do not start camera capture if video i...
Fix #86, do not start camera capture if video is not enabled.  This allows a
user to disable video using either the app options dialog, or in the config
file.  Either way, this allows the webcam to be used by another application
while running microdrop.

Note that the Windows VideoCapture Python library that's currently being used
for the Windows CameraCapture class hangs if an in-use camera ID is specified.
Unfortunately, it seems that this behaviour is in the compiled C extension
portion of the module.  However, this may become a non-issue if/when we switch
to GStreamer for handling the video.