Verify camera capture init with timeout
Verify camera capture init with timeout of 5 seconds.  Unfortunately, although
this works for Ubuntu, Windows still seems to hang when trying to connect to a
camera that is in use by another application.  For now, this can be worked
around by disabling video manually via the app options dialog or in the
microdrop.ini file.  A more permanent solution will hopefully be available if
a switch is made to GStreamer.
See #98.