Add GStreamer libs and SDK to install dependencies
See #108.
Note that the GStreamer libraries must be added to the system PATH
variable and the GStreamer Python bindings must be added to PYTHONPATH.
Also, if using a version of Python other than 2.6, the correct libraries
and Python bindings must be copied into the main GStreamer lib
directory.
Need to perform check in install_dependencies.py to verify these steps
have been done.