Add update_automatically app option to plugin m...
Add update_automatically app option to PluginManagerController.   When
set to True, the plugin server will be be queried and all installed
plugins will be updated if there is a newer version on the plugin
server.  If update_automatically is set to False, then plugins will
only be updated if done so manually in the plugin manager dialog.

Fixes #120.