Fixes #124
Clicking on any of the protocol buttons didn't always trigger events.
Specifically, these events were ignored if the mouse pointer was not
over the button widget at the time that the event was processed. This
commit uses the location of the mouse pointer at the time that the event
was fired.