Remove dependency_links...
Remove `dependency_links` since they cause issues with recent versions
of `pip`, which require the specification of [`allow-external` and
`allow-unverified`][1] for install to work properly for this package.
In an effort to simplify things for the time-being, we simply list the
dependencies and assume that they are installed, or available through
PyPi.

[1]: https://github.com/pypa/pip/issues/1423