Use run_exe for plugin post-install hook
Use `run_exe` function to run plugin post-install hook, which requests
permission to run with elevated privileges on Windows if an error is
encountered running without eleveated privileges.

This is helpful, for example, if MicroDrop is installed to `Program
Files`, but a plugin attempts to install a Python package as a
post-install action.