[FIX] Use currently selected log path for notebook
Prior to this commit, the "Open" and "New" buttons in the experiment log viewer
dialog launched a file browser in the most recent log directory.  Furthermore,
the IPython notebook was launched in the most recent log directory.

As of this commit, the currently selected log path is used instead of the most
recent log directory for all notebook tasks.