Set app.dmf_device before on_dmf_device_swapped
Prior to this commit, if the `microdrop.app` plugin processed the
`on_dmf_device_swapped` signal after the
`wheelerlab.device_info_plugin`, the `get_device` plugin call intended
to notify other plugins that the device has been swapped would refer to
the *previous* DMF device (or `None` if there was no previous device).

As of this commit, the `wheelerlab.device_info_plugin` requests to
schedule the handling of the `on_dmf_device_swapped` signal after the
`microdrop.app` plugin.  This ensures that the correct DMF device is
referenced in the response from the `get_device` plugin call.