[FIX] Multiple electrodes mapped to same channel
Fix handling of setting electrode state where multiple electrodes are
mapped to same channel.

Prior to this commit, setting electrode state in cases where multiple
electrodes are mapped to same channel could result in a `pandas`
indexing error.

As of this commit, setting electrode state where multiple electrodes are
mapped to same channel updates the state of all electrodes for the
affected channel.