[NB] Change behaviour when loading external device
Change behaviour when loading a device file from outside the profile
directory.

If selected device file is not in MicroDrop devices directory, the
device file is copied to devices directory under a subdirectory with
same name as file under the `devices` profile directory.

In addition, the user is informed that device was copied from the
original location into the MicroDrop profile `devices` directory.