Save extracted module names for enabled plugins
Since v2.10.3, Conda package names for enabled plugins were recorded to
the `microdrop.ini` file instead of the corresponding plugin Python module
names.  This caused an issue when attempting to import the plugins in the
saved list of enabled MicroDrop plugins, since the imports should
reference the Python module and **not** the Conda package name.

As of this commit, the plugin module names for enabled plugins are
extracted from the corresponding Conda package names and saved to the
`microdrop.ini` configuration file.