fix(shutdown): process shutdown in main thread
Prior to this commit, application shutdown was scheduled to be executed in
the main GTK thread using `gobject.idle_add`.  However, this didn't work
as expected in cases where the GTK main loop was no longer running.

As of this commit, process shut down directly (instead of using
`gobject.idle_add`) if executing in main thread.