feat(frozen): load icon using `pkgutil`
Prior to this commit, the path to the MicroDrop icon was found by locating
the path relative to the `microdrop` Python module.  However, this
approach failed if the `microdrop` module was [read from within a `.zip`
archive][zip] (e.g., in app packaged with [Py2Exe][py2exe]), since the
icon file path would not be to a regular file, but to a location within
the `.zip` archive.

In this commit, read the icon file using `pkgutil` to support loading from
`.zip` archives.

[zip]: https://docs.python.org/2.3/whatsnew/node5.html
[py2exe]: http://www.py2exe.org