fix(pyutilib): fix PluginGlobals.env accesses
Prior to this commit, `PluginGlobals.env` was accessed as a function,
according to the API for `pyutilib<5.0`.  However, as of `pyutilib>=5.0`,
the API has changed such that `PluginGlobals.env` is a dictionary instead
of a function.

In this commit, replace all `PluginGlobals.env` function calls to
dictionary lookups to support the `pyutilib>=5.0` API.