fix(electrode-ctrl): on_step_swapped threadsafety
Prior to this commit, `on_step_swapped()` used the `plugin` attribute
directly to send a ZeroMQ interface request.  However, the `plugin`
attribute is initialized in a background thread and using it to make
ZeroMQ interface calls is not thread-safe, since the associated ZeroMQ
sockets are not thread-safe.

In this commit, update `on_step_swapped()` to use
`plugin_helpers.hub_execute_async()` to ensure thread-safe calls to the
ZeroMQ interface.