fix(electrode-ctrl): init ZMQ plugin in thread
Prior to this commit, the ZeroMQ plugin instance (along with the
associated ZeroMQ socket instances) were _initialized_ in the main GTK
thread, but were _used_ from the background monitoring thread.  Since
ZeroMQ sockets are not thread-safe, this usage is not thread-safe either.

In this commit, initialize ZeroMQ plugin instance in the background
monitoring thread to ensure thread-safety.