fix(protocol-ctrl): fix on_step_options_changed
Prior to this commit, the protocol controller executed the `run_step()`
method in the `on_step_options_changed` handler.

In this commit, modify `on_step_options_changed()` handler to not execute
`run_step()` and instead only emit `on_step_swapped` iff protocol is not
running.  Doing so avoids trying to run the step when it is already
running.