fix(command-plugin): init plugin in monitor thread
Prior to this commit, the ZeroMQ plugin instance for the `command_plugin`
was initialized in the GTK main thread.  However, since the plugin sockets
were accessed by monitor thread, this might have caused thread-safety
issues.

In this commit, initialize the ZeroMQ plugin instance in the monitor
thread, wrapped in a cancellable task, i.e., with a `cancel()` method to
stop execution.