fix(electrode-ctrl): use duration for 0 electrodes
Prior to this commit, if no electrodes were specified for actuation, a
duration of 0 was used for all actuation requests.  However, there may be
some cases where a user would like to wait for a specified duration with
no electrodes actuated.

In this commit, still apply actuation for specified duration even if _no
electrodes_ are specified for actuation.  If required, zero length
duration can be specified explicitly.