fix(electrode-ctrl): clear dynamic with 0 duration
Prior to this commit, on steps with dynamic actuations, the duration
setting of the active step was used to execute the final actuation of the
step where any dynamic actuations were cleared.  However, this resulted in
an unexpected delay at the end of each such step.

In this commit, on steps with dynamic actuations, set duration to zero for
final iteration to effectively disable previous dynamic actuations before
completing the step.