feat(device): use MICRODROP_DEFAULT_DEVICE env var
Prior to this commit, if a device name was not specified in the MicroDrop
configuration file upon startup, the first available device in the
`devices` directory was used.  However, in some cases, it may be
convenient to explicitly specify the default device that should be used in
such cases.

In this commit, check for the default device setting in the
`MICRODROP_DEFAULT_DEVICE` environment variable.  If set, load the device
matching the specified name upon startup if no device name is specified in
the MicroDrop configuration file.