fix(experiment-log): no new log on app exit
Prior to this commit, a new experiment log was created immediately after
each experiment log was saved.  However, in some cases, e.g., when exiting
the application, it is not necessary to create a new experiment log.

In this commit, add an optional `create_new` parameter to the `save()`
method.  If `False`, do not create a new experiment log after saving the
current log.  Set `create_new=False` when saving the experiment log on
application exit.