fix(electrode-ctrl): ignorable actuation errors
Prior to this commit, if an electrode actuator (i.e., a plugin
implementing the `IElectrodeActuator` interface) raised an exception
during a `on_actuation_request()` call, the exception was raised,
preventing further actuations.  However, in some cases the user might want
acknowledge the error and continue anyway, e.g., when a DropBot is not
connected but the plugin is still enabled.

In this commit, display a warning prompt to the user when a actuation
fails, but give the user the option to ignore the error and continue.

Fixes #261.