fix(electrode-ctrl): use default step kwargs
Prior to this commit, if no kwargs were specified for the electrode
controller plugin, an exception would occur during step execution.  This
happened, for example, when creating a new protocol while real-time mode
was enabled since the only step in the new protocol starts empty without
_any_ plugin kwargs.

In this commit, if no kwargs are specified for the electrode controller,
use the default step options.  This is preferable to simply skipping step
execution in the electrode controller since it ensures that electrode
actuators are updated according to the default electrode controller step
settings, including **all electrodes turned off**, default voltage, etc.