feat(prompt-plugin): add prompt plugin
Add `microdrop.prompt_plugin` core plugin to handle prompting user for
input.  The goal of this plugin is to decouple other plugins from GTK GUI
code.

In this commit, implement a receiver for the `'warning'` signal in the
step `signals` that displays a warning prompt.  The user is given the
option of ignoring the warning and continuing.  If the user chooses to not
ignore the warning, an exception is raised.