feat(electrode-ctrl): add warning signal wrapper
Add `_warning()` function to simplify sending `'warning'` signal.

As of this commit, the `microdrop.prompt_plugin` is the only plugin
registered as a receiver of the `'warning'` signal.  When a `'warning'`
signal is received by the `microdrop.prompt_plugin`, a warning dialog is
displayed and the user is given the option to ignore the warning an
continue.  If the user chooses to not ignore the warning, an exception is
raised.