refactor(electrode-ctrl): add static_states param
Prior to this commit, the `execute_actuations()` method looked up the
static electrode states during step execution in the step options via the
`StepOptionsController` API.  However, this violated the goal of having
all step execution parameters enclosed in the `plugin_kwargs` passed to
`on_step_run()`.

In this commit, retrieve the static electrode states in the
`plugin_kwargs` passed to `on_step_run()`.