refactor(electrode-ctrl): isolate pyutilib code
Prior to this commit, the electrode controller Pyutilib plugin code was
located in the root module of the plugin directory.  However, this
prevented importing the `execute` module without also importing the
Pyutilib plugin.

In this commit, refactor the electrode controller plugin to move the
Pyutilib code into a submodule named `pyutilib`.