feat(plugins): import from pyutilib if available
Prior to this commit, Pyutilib plugins were only imported directly from
`__init__.py` within a plugin directory.  However, in some cases it may be
beneficial to import submodules of a plugin _without_ implicitly importing
and initializing the corresponding Pyutilib plugin.

In this commit, support isolating the Pyutilib plugin code within a
`pyutilib` submodule (i.e., `pyutilib.py`) within the plugin.  This is
done by importing from `pyutilib` submodule in plugin, if it exists,
instead of the importing the root of the plugin.