chore(main-window): rearrange status labels
Prior to this commit, status labels in the main window (e.g., device,
protocol, experiment, etc.) were placed directly in the same `gtk.VBox` as
the protocol controls (e.g., previous, play, next) and the protocol
controller.  This prevented any other widgets from being placed _beside_
the status labels to utilize the empty space in the window.

In this commit, rearrange the status labels by moving them into a
`gtk.VBox` within a `gtk.HBox`.  This makes it possible for other plugins
to add widgets to the `gtk.HBox` as required.