feat(default-dirs): move protocols out of devices
Prior to this commit, each device layout required its own directory within
the default `devices` directory.  A `protocols` directory was then created
within each device layout directory to store protocols relevant to that
device.  Since the locations of devices and protocols were fixed, a simple
text-entry dialog was displayed when saving protocols or devices,
prompting only for a `name` to save, since the destination directory was
implied by the directory layout convention.

However, this approach had some disadvantages:

 - User has no control over where a device or protocol is saved.
 - The save location is not transparent to the user.
 - A protocol is not necessarily only relevant to a single device
   (actuation states are specified in terms of electrode IDs, which may be
   compatible between device layouts).

In this commit:

 - do not create a directory for each device; instead use the base of the
   SVG filename as the device name directly
 - change the default location for `protocols` directory to sit next to
   the default `devices` directory
 - present a standard file browser to select a location for saving
   protocols or devices.

As part of these changes, the full file path of the most recently used
protocol/device is written to the config file rather than just the _name_.