fix(experiment-log): save if log dir not empty
Prior to this commit, the experiment directory was only saved if one or
more plugins had explicitly added data to the global `app.experiment_log`
instance of the `ExperimentLog` class.  However, adding data to the
`ExperimentLog` object has been deprecated and is no longer used by
plugins.

In this commit, instead save the experiment directory if it is not empty
(i.e., contains at least one file or directory).