feat(protocol): use PROTOCOLS_DIR for load/save
Prior to this commit, the default directory used when saving or loading
protocols was inferred as a directory named `protocols` under the same
parent directory as the default `devices` directory.  However, the default
protocols directory defined as `PROTOCOLS_DIR` in the
`microdrop.default_paths` module should be used instead.

In this commit, use the `PROTOCOLS_DIR` path from the
`microdrop.default_paths` module as the default directory when saving or
loading protocols.