chore(protocol-ctrl): save absolute path to config
Prior to this commit, if a loaded protocol path was within the default
protocols directory, a relative path to the protocol was stored in the
config file as the most-recently used protocol.

In this commit, always store the _absolute path_ to a loaded protocol in
the config file, regardless of whether or not the protocol is located in
the default protocols directory.