feat(device-connections): prompt to save as new
Prior to this commit, connections detected in response to the
`Device > Detect _connections...` were forcefully written to the original
device file, overwriting the file without prompting the user for
confirmation.

In this commit, use a _"Save as"_ dialog to prompt the user to enter an
output location for the modified device file (user may explicitly choose
to overwrite the original file path, if desired).  Also, display dialog
while detecting connections to indicate progress.