feat(experiment): continue dialog after protocol
Prior to this commit, an experiment log was only saved/archived if a new
experiment was explicitly requested through the `File > New experiment`
menu item.

In this commit, offer to create a new experiment if experiment data exists
after protocol has completed, giving the user the option to either
continue the current experiment or create a new one.