fix(protocol): filename as label in main window
Prior to this commit, the `name` attribute of the loaded protocol was
displayed as the protocol name in the main application window.  However,
use of the `name` attribute has been deprecated in all other code and the
`name` attribute will be removed in a future release.

In this commit, use the protocol _filename_ as the displayed name in the
main window label.